Add reducer for greetings
