+++ app/javascript/api/fetchAPI.js
@@ -1,10 +1,10 @@
 const fetchAPI = async () => {
-  const URL = 'http://127.0.0.1:3000/api/v1/messages';
+  const URL = 'http://127.0.0.1:3000/v1/messages';
   const request = await fetch(URL, {
     method: 'GET',
     headers: {
-      Accept: 'application/json'
-    }
+      Accept: 'application/json',
+    },
   });
   const response = await request.json();
   return response;
