--- app/javascript/components/NoMatch.js
@@ -1,9 +1,0 @@
-import React from "react"
-
-const NoMatch = () => (
-  <div className="container load">
-    <h1>Sorry Page not Found!</h1>
-  </div>
-);
-
-export default NoMatch;
